Add bash and zsh completions for llppac
Use two separate lists (llpp.in, llppac.in) of file extensions and
templates ({bash,zsh}/{llpp,llppac}.mk) to generate completions for both
executables and both shells.  The script mkcomp.sh facilitates this.

Added option completion for llpp and llppac in zsh

Employ a Makefile for building/installing the files