Move cursor handling to C
Theming and such is handled outside of the protocol (in the Xlib).

Default X11 font cursors are somewhat unpretty, but here custom cursor
font was used and it was never much of a bother, things change though,
let's follow suit.

On one hand healthy -~30 lines, on the other Xlib... oh well, given
it's inevitable presence due to GLX...