Manual buffer clears
On this hardware/software combination glClear is evidently rather
power taxing. The effect is not noticeable when measuring system load
in the usual way but RAPL readings (and/or external wattmeter) tell
different story - glClear is not free. More investigation is required
though, for now just clear parts of the buffer manually, and only in
single column mode, hopefully "not this" systems wouldn't suffer.