Add anki deck to documentation
This is a suggestion to help users with remembering the not often used yet useful shortcuts.
Having to search the help text everytime we need something is time consuming, and we might forget that a feature is there at all (eg `[` and `]` for brightness control. It's simpler to just learn them :)