Bump memlimit
(Previous): 32M was born or a mac mini ppc with 128M of RAM (and 32M
of dedicated VRAM) and a CRT sporting 1280x1024 resolution at 16bpp,
(current) 128M is a personal setting on an intel nuc with 8G (shared)
and an LCD with 3840x2160/32bpp. Decade(s) late welcome to the 21st
century i suppose.