Remove remapping help/hist hack
Some keyboards (like the one that is used to write this commit
message) do not have arrow keys, thus remappping code/hack was
introduced that maps home row key_codes_ to arrows. h (help/history)
is on the home row (Dvorak) and so the hack on the hack was added to
recognize those as just letters when certain masks are used. But
alt-left/right was reserved for navigational history traversal. So
remove the second hack layer and deal with the fact that history is
not accessible in a remapped case. The whole situation is less than
ideal.