Papper over #160
PDF attached to the bug report causes MuPDF to emit a lot of
diagnostic messages like
  warning: FT_Get_Advance(HiddenHorzOCR,32534): invalid glyph index

This happens inside ml_whatsunder which calls emsg an the same thread
that drains the socket thus causing a deadlock. The solution is either
calling back OCaml from C, not a fan, or doing nothing, which is the
current "solution".