Change getlinkn
Negative and zero indicate the fact that link wasn't found, otherwise
one plus link index is returned. Price one has to pay for not wanting
to deal with exceptions or sum type returns.

This fixes the behavior in cases where a link is on the page following
the one without links at all.