Hack around intermittent postprocess failures
How it was found:

An attempt was made to search for 'a' in an infinitely tall document
(HTML font test page produced by llppac with
  Extended parameters » reflowable Layout » height
 set to zero)

Searching took surprisingly long time. At some point ml_postprocess
was trying to acquire the lock, failed, schedulled a redisplay, which
failed because concurently running earch was holding it), tho "fix" is
to not even attempt postprocessing.

Maybe someday enough courage will be found to tackle the problem
"properly".