Remove bogus(?) assert
01ff5cb4ce0471bd4974e60d8a8eaaf4ce71d994 commented out redisplay and
added assert, without specifying why, will err on the side of "it was
a debugging leftover" (after all the commit in question was a WIP) and
return things to the way they were (i.e. post redisplay instead of
dying)

Currently the assert can be triggered while navigating ML2015-talk.pdf
in presentation mode.